Training Projects/Ollama News-Summarizer/source/requirements.txt
feedparser
requests
streamlit
python-dateutil
nltk
geocoder
pytz
langdetect
openai
newspaper3k
urllib3
